# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mango",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/assinaturas": {
      "get": {
        "tags": [
          "assinatura"
        ],
        "summary": "Listar",
        "operationId": "listar_api_assinaturas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssinaturaRead"
                  },
                  "type": "array",
                  "title": "Response Listar Api Assinaturas Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "assinatura"
        ],
        "summary": "Criar",
        "operationId": "criar_api_assinaturas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssinaturaCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssinaturaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assinaturas/{item_id}": {
      "get": {
        "tags": [
          "assinatura"
        ],
        "summary": "Obter",
        "operationId": "obter_api_assinaturas__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssinaturaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "assinatura"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_assinaturas__item_id__patch",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssinaturaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssinaturaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "assinatura"
        ],
        "summary": "Remover",
        "operationId": "remover_api_assinaturas__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/recuperar-senha": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Recuperar Senha",
        "operationId": "recuperar_senha_api_auth_recuperar_senha_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecuperarSenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contas/saldos-diarios": {
      "get": {
        "tags": [
          "conta"
        ],
        "summary": "Saldos Diarios",
        "operationId": "saldos_diarios_api_contas_saldos_diarios_get",
        "parameters": [
          {
            "name": "dias",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 30,
              "title": "Dias"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContaSaldoSerie"
                  },
                  "title": "Response Saldos Diarios Api Contas Saldos Diarios Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/contas/{conta_id}": {
      "get": {
        "tags": [
          "conta"
        ],
        "summary": "Obter",
        "operationId": "obter_api_contas__conta_id__get",
        "parameters": [
          {
            "name": "conta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conta Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContaDetalheRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "conta"
        ],
        "summary": "Vincular Objetivo",
        "operationId": "vincular_objetivo_api_contas__conta_id__patch",
        "parameters": [
          {
            "name": "conta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conta Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/contas/{conta_id}/faturas-resumo": {
      "get": {
        "tags": [
          "conta"
        ],
        "summary": "Faturas Resumo",
        "operationId": "faturas_resumo_api_contas__conta_id__faturas_resumo_get",
        "parameters": [
          {
            "name": "conta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conta Id"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 24,
              "minimum": 1,
              "default": 6,
              "title": "Limite"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaturasResumo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/contas/{conta_id}/instituicao": {
      "put": {
        "tags": [
          "conta"
        ],
        "summary": "Vincular Instituicao",
        "operationId": "vincular_instituicao_api_contas__conta_id__instituicao_put",
        "parameters": [
          {
            "name": "conta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conta Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContaInstituicaoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssinaturaCreate": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "valor_centavos": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Valor Centavos"
          },
          "categoria_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria Id"
          },
          "periodicidade": {
            "type": "string",
            "enum": [
              "mensal",
              "trimestral",
              "semestral",
              "anual",
              "irregular"
            ],
            "title": "Periodicidade"
          },
          "data_inicio": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Inicio"
          },
          "ativa": {
            "type": "boolean",
            "title": "Ativa",
            "default": true
          },
          "detectada_automaticamente": {
            "type": "boolean",
            "title": "Detectada Automaticamente",
            "default": false
          },
          "conta_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conta Id"
          },
          "nomes_transacao": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Nomes Transacao"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "valor_centavos",
          "periodicidade"
        ],
        "title": "AssinaturaCreate"
      },
      "AssinaturaRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "nomes_transacao": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Nomes Transacao"
          },
          "valor_centavos": {
            "type": "integer",
            "title": "Valor Centavos"
          },
          "categoria_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria Id"
          },
          "periodicidade": {
            "type": "string",
            "title": "Periodicidade"
          },
          "data_inicio": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Inicio"
          },
          "ativa": {
            "type": "boolean",
            "title": "Ativa"
          },
          "detectada_automaticamente": {
            "type": "boolean",
            "title": "Detectada Automaticamente"
          },
          "conta_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conta Id"
          },
          "usuario_id": {
            "type": "integer",
            "title": "Usuario Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "nomes_transacao",
          "valor_centavos",
          "periodicidade",
          "ativa",
          "detectada_automaticamente",
          "usuario_id",
          "criado_em",
          "atualizado_em"
        ],
        "title": "AssinaturaRead"
      },
...
      "AssinaturaUpdate": {
        "properties": {
          "nome": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "valor_centavos": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor Centavos"
          },
          "categoria_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria Id"
          },
          "periodicidade": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "mensal",
                  "trimestral",
                  "semestral",
                  "anual",
                  "irregular"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodicidade"
          },
          "data_inicio": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Inicio"
          },
          "ativa": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ativa"
          },
          "nomes_transacao": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nomes Transacao"
          }
        },
        "type": "object",
        "title": "AssinaturaUpdate"
      },
...
      "ContaInstituicaoUpdate": {
        "properties": {
          "pluggy_connector_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pluggy Connector Id"
          },
          "nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          }
        },
        "type": "object",
        "title": "ContaInstituicaoUpdate",
        "description": "Vínculo manual da conta a uma instituição escolhida no catálogo do Pluggy.\n\n`pluggy_connector_id=None` remove o vínculo. Com connector, `nome` é obrigatório e\n`logo_url` é o `imageUrl` do connector (opcional)."
      },
...
      "ContaSaldoSerie": {
        "properties": {
          "conta_id": {
            "type": "integer",
            "title": "Conta Id"
          },
          "pontos": {
            "items": {
              "$ref": "#/components/schemas/SaldoDiarioPonto"
            },
            "type": "array",
            "title": "Pontos"
          }
        },
        "type": "object",
        "required": [
          "conta_id",
          "pontos"
        ],
        "title": "ContaSaldoSerie",
        "description": "Série de saldo diário (fecho de cada dia) de uma conta BANK, para o sparkline dos cards."
      },
      "ContaUpdate": {
        "properties": {
          "objetivo_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objetivo Id"
          }
        },
        "type": "object",
        "title": "ContaUpdate",
        "description": "Update estreito (§4 crud.md): o usuário só altera o vínculo com objetivo.\n\n`objetivo_id=None` desvincula. Campos importados do Pluggy não são editáveis aqui."
      },
...
      "FaturasResumo": {
        "properties": {
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/FaturaResumoBucket"
            },
            "type": "array",
            "title": "Buckets"
          }
        },
        "type": "object",
        "required": [
          "buckets"
        ],
        "title": "FaturasResumo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "title": "Senha"
          },
          "codigo_totp": {
            "type": "string",
            "maxLength": 8,
            "minLength": 6,
            "title": "Codigo Totp"
          }
        },
        "type": "object",
        "required": [
          "email",
          "senha",
          "codigo_totp"
        ],
        "title": "LoginRequest"
      },
      "MeRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "accent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "email"
        ],
        "title": "MeRead"
      },
...
      "RecuperarSenhaRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "codigo_totp": {
            "type": "string",
            "maxLength": 8,
            "minLength": 6,
            "title": "Codigo Totp"
          },
          "nova_senha": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 8,
            "title": "Nova Senha"
          }
        },
        "type": "object",
        "required": [
          "email",
          "codigo_totp",
          "nova_senha"
        ],
        "title": "RecuperarSenhaRequest"
      },
...
      "SaldoDiarioPonto": {
        "properties": {
          "data": {
            "type": "string",
            "format": "date",
            "title": "Data"
          },
          "saldo_centavos": {
            "type": "integer",
            "title": "Saldo Centavos"
          }
        },
        "type": "object",
        "required": [
          "data",
          "saldo_centavos"
        ],
        "title": "SaldoDiarioPonto"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}